Initial commit
Dear Alvaro, I'm writing this git history as a little story about TDD
as I understand it (dissidents!) using the very trite Bowling kata.

For me, TDD is a self-restraint mechanism. When presented with a
problem, many of us, programmers with a little bit of experience and
many patterns and mechanisms to throw at problems it is very easy to
over-engineer anything. Usually, well before reading the complete
problem statement I have a lot of cool ideas about how to solve it
(chain of responsibility, flyweight, multiple loan patterns...) and
sometimes is what you need BUT...

So step 1 in TDD is thinking about the problem and solution space and
having an idea of how the solution might be structured. In the case of
the bowling kata there is a tension between the concepts of roll and
frame in order to count the final score so a nice idea will be to have
a series of frames collaborating in a chain of responsibility. Let's try
to move in that direction... step by step.

First, let's have a project with `sbt new sortega/scala-seed` or any
other mean.