GREEN: trolling myself
Yep. This might be the simplest one passing the test... but not what you
expected. I could have gone right to the `rolls.sum` you were expecting
but this way I can introduce the concept of "triangulation."

If you want to locate the origin of an earthquake you need more than one
seismograph. With just one, there are many combinations of distance and
force that match with the data. If you add a second one, you can reduce
the options, but with a third one you can pinpoint the origin and
magnitude. That is triangulation.

Look at the tests we have. Why not hardcoding 0 and 3 depending on the
first roll? Why not just adding the two first rolls? We need more
assertions to "triangulate".