Still RED: refactor to scale local optimum
Going for the `sum` thrown us into a corner. We need to make decisions
while the sequence of rolls is iterated. So the first step in the
direction of getting green is doing a refactor.

In this case we want to iterate (recursively but choose your own style)
frame by frame through the list. Note that the previous tests are still
green but the new one is still red.