WHAT IF...
In this final commit I'm adding what I would have written if I went
straight to pass all the tests without the step-by-step process of TDD.

This solution introduces case classes for the types of frames, something
we didn't need doing TDD. Apart from that, when "jumping over" TDD steps
it is easier to use higher level powerful functions like `unfold` and
`unfoldWithIndex` that unfortunately are not in the scala library and
I needed to quickly add.

The result is a little bit more complex but that is because I know this
problem very well at this point. For a new problem I expect to have a
lot more accidental complexity when doing free style. In general, you
should indulge yourself in leaving the TDD discipline when there is low
risk of over-engineering or when you know really well the problem
domain.